==> 02_lmpc-infection-rnaseq-analysis-pre-setrank.R <==
################################## Script start
print("Starting script 2")

################################## Set seed for reproducibility
set.seed(1337)

################################## Install packages
renv::restore()

################################## Load packages
lapply(
  c(
    "renv", # For project management
    "BiocManager", # For project management
    "plyr", # Data wrangling, part of tidyverse but not automatically loaded with it. Always load plyr before dply to avoid known issues # nolint: error. # nolint
    "ggplot2", # Tidyverse. Data wrangling, processing and presentation.
    "dplyr", # Tidyverse. Data wrangling, processing and presentation.
    "tidyr", # Tidyverse. Data wrangling, processing and presentation.
    "readr", # Tidyverse. Data wrangling, processing and presentation.
    "purrr", # Tidyverse. Data wrangling, processing and presentation.
    "tibble", # Tidyverse. Data wrangling, processing and presentation.
    "stringr", # Tidyverse. Data wrangling, processing and presentation.
    "forcats", # Tidyverse. Data wrangling, processing and presentation.
    "vroom", # Faster data wrangling
    "lubridate", # Working with dates, part of tidyverse but not automatically loaded with it
    "svglite", # To make svg files with ggsave
    "writexl", # Writing excel files
    "DESeq2", # Differential gene expression
    "IHW", # Better power for adjusting p-values of differential gene expression
    "tximport", # Importing RNAseq pipeline data
    "tximportData", # Importing RNAseq pipeline data
    "umap", # For umap, Uniform Manifold Approximation and Projection
    "EnsDb.Mmusculus.v79", # ensdb package for mouse
    "ensembldb", # For getting gene lengths necessary for TPM calculation
    "SetRank", # For the SetRank GSEA
    "parallel", # For parallelisation
    "biomaRt", # For annotation and GO gene sets
    "org.Mm.eg.db", # For GO term annotation, might be used instead of biomaRt
    "reactome.db", # For annotationdbi of reactome
    "GO.db", # For GO term annotation, might be used instead of biomaRt
    "KEGGREST", # For KEGG
    "styler", # R studio addin for interactively adhere to the tidyverse style guide
    "RCy3", # For cytoscape programmatic analysis
    "STRINGdb", # For STRING database annotation
    "igraph", # For RCy3/cytoscape
    "viridis" # Color palette
  ),
  library,
  character.only = TRUE
)

################################## Check existance of required directories
if (!dir.exists("P26010")) {
  stop("You need the P26010 root folder in the same directory as the script")
} else {
  print("./P26010 exists")
}

if (!dir.exists("R_input_files")) {
  stop("You need the R_input_files root folder in the same directory as the script")
} else {
  print("./R_input_files exists")
}

################################## Create R intermediate directory if it doesn't exist
if (!dir.exists("R_intermediate_files")) {
  dir.create("R_intermediate_files")
  print("Creating directory ./R_intermediate_files")
} else {
  print("./R_intermediate_files exists")
}

################################## Create R output directory if it doesn't exist
if (!dir.exists("R_output_files")) {
  dir.create("R_output_files")
  print("Creating directory ./R_output_files")
} else {
  print("./R_output_files exists")
}

################################## Create SetRank results subdirectory if it doesn't exist
if (!dir.exists("./R_output_files/Setrank_results")) {
  dir.create("./R_output_files/Setrank_results")
  print("Creating directory ./R_output_files/Setrank_results")
} else {
  print("./R_output_files/Setrank_results exists")
}

################################## Create plots results subdirectory if it doesn't exist
if (!dir.exists("./R_output_files/Figures")) {
  dir.create("./R_output_files/Figures")
  print("Creating directory ./R_output_files/Figures")
} else {
  print("./R_output_files/Figures exists")
}

################################## Create tables results subdirectory if it doesn't exist
if (!dir.exists("./R_output_files/Tables")) {
  dir.create("./R_output_files/Tables")
  print("Creating directory ./R_output_files/Tables")
} else {
  print("./R_output_files/Tables exists")
}

################################## Data import
# All file reading assumes you are in a root directory above P26010
# Import sample info file from 00-Reports, annotate sample list and convert potentially problematic characters in column names
# Change path to where in your directory the file S.Linden_22_01_sample_info.txt is
# Excluding samples from treatment group as they are outside the scope of this article
# Creating new set of IDs for manuscript that is a bit more intuitive
replacement_list_User_ID <- c("M-1" = "H9_01",
                                "M-2" = "H9_02",
                                "M-3" = "H9_03",
                                "M-4" = "H9_04",
                                "M-5" = "H9_05",
                                "M-6" = "H9_06",
                                "M-7" = "H9_07",
                                "M-8" = "H9_08",
                                "M-9" = "H9_09",
                                "M-10" = "H9_10")

Sample_List <- read_tsv(
  file = "./P26010/00-Reports/S.Linden_22_01_sample_info.txt",
  col_types = c("ccdd")
) %>%
  dplyr::rename(names = "NGI ID") %>%
  dplyr::rename(User_ID = "User ID") %>%
  dplyr::mutate(Manuscript_ID = str_replace_all(User_ID, setNames(names(replacement_list_User_ID), replacement_list_User_ID))) %>% 
  dplyr::filter(!User_ID %in% c(
    "H9_11",
    "H9_12",
    "H9_13",
    "H9_14",
    "H9_15"
  )) %>%
  mutate(Infected = c(
    "Non_Infected", "Non_Infected", "Non_Infected", "Non_Infected", "Non_Infected",
    "Infected", "Infected", "Infected", "Infected"
  )) %>%
  mutate(condition = Infected)

# Importing data about the LMPC and the RNA purification
# Excluding samples from treatment group since they are outside the scope of this article
# Code assumes this data file is in the same directory as the script
LMPC_RNA_Data <- read_tsv(
  file = "./R_input_files/LMPC_RNA_Data.txt",
  col_types = c("cifdddccidddd")
) %>%
  mutate("Date_harvested" = lubridate::as_date(Date_harvested)) %>% # I couldn't get the date to parse correctly with readr so I change from character to date after reading
  mutate("Date_sectioned" = lubridate::as_date(Date_sectioned)) %>%
  dplyr::filter(!User_ID %in% c(
    "H9_11",
    "H9_12",
    "H9_13",
    "H9_14",
    "H9_15"
  ))

files <- file.path(
  "./P26010/01-RNA-Results/star_salmon",
  Sample_List$names,
  "quant.sf"
)
print(files)
file.exists(files)

# tx2gene from the rnaseq pipeline
tx2gene <- read_tsv(
  file = "./P26010/01-RNA-Results/genome/salmon_tx2gene.tsv",
  col_names = c(
    "TXNAME",
    "GENEID",
    "GeneSymbol"
  ),
  col_types = (c("ccc"))
)

# import star salmon results with tximport
txi <- tximport(files, type = "salmon", tx2gene = tx2gene)
glimpse(txi$counts)
colnames(txi$counts) <- Sample_List$names
glimpse(txi$counts)

################################## Investigate statistics about sequencing depth
# Mean sequencing depth and standard deviation
Sample_List %>%
  dplyr::summarise("Mean sequencing depth in mreads" = mean(Mreads))

Sample_List %>%
  dplyr::summarise("Sequenced reads standard deviation" = sd(Mreads))

################################## Investigate statistics about LMPC and RNA purification data
## Table of statistics about sectioning time, LMPC area, RIN and V200 values
Summary_LMPC_RNA_data <- LMPC_RNA_Data %>%
  dplyr::select(
    Days_between_harvest_and_sectioning,
    Extracted_area_with_10_micrometer_thick_sections,
    Tapestation_RIN_Value,
    SciLifeLab_RIN_Value,
    DV200,
    Tapestation_RNA_concentration_ng_ul,
    SciLifeLab_RNA_concentration_ng_ul,
    SciLifeLab_RNA_amount_ng
  ) %>%
  rename_all(~ str_replace_all(., "_", "-")) %>%
  dplyr::summarise(across(everything(), list(mean = mean, median = median, min = min, max = max))) %>%
  pivot_longer(cols = everything(), names_to = "summary_stat", values_to = "value") %>%
  separate(summary_stat, into = c("Variable", "summary_stat"), sep = "_") %>%
  pivot_wider(names_from = summary_stat, values_from = value)

# Write table of summary statistics
write_xlsx(
  x = Summary_LMPC_RNA_data,
  path = "./R_output_files/Tables/Summary_LMPC_RNA_data.xlsx"
)
Summary_LMPC_RNA_data

## Any correlation between extracted area and purified RNA?
# Fit a linear model
lm_model <- lm(SciLifeLab_RNA_amount_ng ~ Extracted_area_with_10_micrometer_thick_sections, data = LMPC_RNA_Data)

# Get the model summary
summary_lm <- summary(lm_model)

# Extract the R-squared value
r_squared <- summary_lm$r.squared

# Print the R-squared value
print(r_squared)

# Visualise correlation
LMPC_RNA_Plot <- ggplot(
  LMPC_RNA_Data,
  aes(
    x = Extracted_area_with_10_micrometer_thick_sections,
    y = SciLifeLab_RNA_amount_ng,
    color = condition,
    shape = Sequenced
  )
) +
  geom_point() +
  scale_color_manual(values = c("Non_Infected" = "#4BC26C", "Infected" = "#440D54")) +
  geom_hline(yintercept = 10) +
  geom_smooth(method = "lm", se = FALSE, color = "red") +
  expand_limits(y = 0, x = 0) +
  scale_y_continuous(breaks = seq(0, max(LMPC_RNA_Data$SciLifeLab_RNA_amount_ng), by = 10)) +
  scale_x_continuous(breaks = seq(0, max(LMPC_RNA_Data$Extracted_area_with_10_micrometer_thick_sections), by = 1000000)) +
  labs(
    title = paste("R-squared =", round(r_squared, 3)),
    x = "LMPC catapulted area (square micrometers)",
    y = "Purified RNA (nanograms)"
  ) +
  theme_minimal()
LMPC_RNA_Plot
ggsave(
  filename = "./R_output_files/Figures/LMPC_RNA_Plot.eps",
  plot = LMPC_RNA_Plot,
  width = 990,
  height = 1550,
  units = "px"
)

################################## Create DESeq object
### There are ENSEMBL IDs which are matched to the same gene symbol
### Our raw data have to be imported with Tximport the best way
### The best way I could think of to handle the duplicated symbols is a multi-step process:
### 1 Create DESeq object from tximport.
### 2 Extract counts from DESeq object
### 3 Merge duplicated GeneSymbols
### 4 Re-create DESeq object with the merged counts

## 1 Create DESeq object from tximport.
# Sample annotation data frame for deseq2
sampleTable <- data.frame(
  condition = factor(
    x = Sample_List$condition,
    levels = c(
      "Non_Infected",
      "Infected"
    )
  ),
  row.names = Sample_List$names
)

# Create table matching ENSEMBL gene id to gene symbol for future annotation
Gene_Symbols <- tx2gene %>%
  dplyr::select(-TXNAME) %>%
  distinct() %>%
  arrange(GENEID)

# Prepare deseq2 object from tximport, this is the "original counts and offset" from the vignette
txi_dds <- DESeqDataSetFromTximport(txi, sampleTable, ~condition)

## Extract counts from DESeq object
Counts_All <- data.frame(BiocGenerics::counts(txi_dds)) %>%
  rownames_to_column("GENEID") %>%
  inner_join(Gene_Symbols, # With full join it will be a complete match with tx2gene, but there will be NA values
    by = "GENEID"
  ) %>%
  replace(is.na(.), 0) # Replace NA values with 0

# Merge counts/gene ids for genes which have been annotated with the same symbol
# Every gene should only be matched to one gene symbol, otherwise might become a problem in downstream analysis
# First check if there are any gene gene symbols mapped to multiple gene IDs
Counts_All %>%
  dplyr::select(GENEID, GeneSymbol) %>%
  distinct() %>%
  plyr::count("GeneSymbol") %>% # this is not a typo, it uses the package plyr and not dplyr
  dplyr::filter(freq > 1)
# Yes. mmu-mir-669a-4 even has 9.
# As sanity check since this shouldn't be possible, are there any gene ids mapped to multiple symbols?
Counts_All %>%
  dplyr::select(GENEID, GeneSymbol) %>%
  distinct() %>%
  plyr::count("GENEID") %>% # sic!
  dplyr::filter(freq > 1)
# No

## Merge duplicated GeneSymbols. Approach is to simply take the first of GENEIDs mapped to the symbol.
Counts_All_NoDuplicatedSymbol <- Counts_All %>%
  pivot_longer(
    cols = starts_with("P26010"),
    names_to = "names",
    values_to = "Counts"
  ) %>%
  group_by(GeneSymbol, names) %>%
  dplyr::summarise("Counts" = sum(Counts)) %>%
  # ddply(.(GeneSymbol, names), summarise, "Counts" = sum(Counts)) %>%
  pivot_wider(
    values_from = Counts,
    names_from = names
  ) %>%
  inner_join(Gene_Symbols,
    by = "GeneSymbol",
    multiple = "first" # this takes the first of the multiple matches. Might not work depnding on package verison
  ) %>%
  distinct(GeneSymbol, .keep_all = TRUE) # If the mutiple = "first" doesn't work with your R version this should work
# Double check duplicates were removed
Counts_All_NoDuplicatedSymbol %>%
  dplyr::select(GENEID, GeneSymbol) %>%
  distinct() %>%
  plyr::count("GeneSymbol") %>% # sic!
  dplyr::filter(freq > 1)
# They were removed

# Convert to tidy format (easier to work with)
Tidy_Counts_All_NoDuplicatedSymbol <- Counts_All_NoDuplicatedSymbol %>%
  pivot_longer(
    cols = starts_with("P26010"),
    names_to = "names",
    values_to = "Counts"
  )

# Add variable showing which genes have a count of at least 1 in at least 1 sample ("TRUE"), and which ones we did not detect ("FALSE")
Tidy_Counts_DetectedGenes_NoDuplicatedSymbol <- Tidy_Counts_All_NoDuplicatedSymbol %>%
  ddply(.(GENEID), summarise, "Summed_Counts" = sum(Counts)) %>%
  mutate("Gene_Detected" = ifelse(Summed_Counts > 0, TRUE, FALSE)) %>%
  dplyr::select(-Summed_Counts) %>%
  inner_join(Tidy_Counts_All_NoDuplicatedSymbol,
    by = "GENEID",
    multiple = "all"
  ) %>%
  inner_join(Sample_List %>% dplyr::select(names, User_ID),
    by = "names"
  ) %>%
  dplyr::select(-names)

## 4 Re-create DESeq object with the merged counts
# Now matrix to put back into deseq2
# Since it is easier to understand gene symbols and we now have unique symbols, we will use that primarily
Counts_Matrix_Unique_Symbols <- as.matrix(Tidy_Counts_DetectedGenes_NoDuplicatedSymbol %>%
  pivot_wider(
    values_from = Counts,
    names_from = User_ID
  ) %>%
  dplyr::select(-GENEID, -Gene_Detected) %>%
  column_to_rownames("GeneSymbol"))

mtx_dds <- DESeqDataSetFromMatrix(
  countData = Counts_Matrix_Unique_Symbols,
  colData = Sample_List %>%
    dplyr::select(User_ID, condition) %>%
    mutate(condition = factor(condition,
      levels = c(
        "Non_Infected",
        "Infected"
      )
    )) %>%
    column_to_rownames("User_ID"),
  design = ~condition
)

## Process with DESeq2
dds <- DESeq(mtx_dds)

################################## PCA and Uniform Manifold Approximation and Projection (umap) of gene expression
## Using principal components as input for umap

## PCA
## Principal Component Analysis(PCA)
# PCA PC1 and PC2
rld <- DESeq2::rlog(dds, blind = TRUE)
rld_mat <- SummarizedExperiment::assay(rld)
pca <- prcomp(t(rld_mat))
df <- cbind(Sample_List, pca$x)

PCA <- ggplot(df, aes(
  x = PC1,
  y = PC2,
  color = condition
)) +
  scale_colour_manual(values = c("Non_Infected" = "#4DC36B", "Infected" = "#440C55")) +
  geom_label(aes(label = Manuscript_ID),
             size = 2) +
  labs(
    x = paste0(
      "PC1 (captures ",
      summary(pca)$importance[2, 1],
      " of variance)"
    ),
    y = paste0(
      "PC2 (captures ",
      summary(pca)$importance[2, 2],
      " of variance)"
    ),
    title = paste0(
      "PC1 and PC2 together captures ",
      summary(pca)$importance[3, 2],
      " of variance in gene expresssion"
    )
  ) +
  theme_bw()
print(PCA)

ggsave(
  filename = "./R_output_files/Figures/PCA.eps",
  plot = PCA,
  width = 1484,
  height = 700,
  units = "px"
)

## Visualise how much each PC explains variance in order to select PC for UMAP
summary_pca <- summary(pca)
PC_Variance <- ggplot(
  enframe(summary_pca$importance[2, ],
    name = "PC",
    value = "Proportion_of_variance_captured"
  ) %>%
    mutate(PC = str_replace(PC, "(^[:alpha:]{2})(\\d)$", "\\10\\2")),
  aes(
    x = PC,
    y = Proportion_of_variance_captured
  )
) +
  geom_point() +
  theme_bw()
print(PC_Variance)

ggsave(
  filename = "./R_output_files/Figures/PC_Variance.eps",
  plot = PC_Variance,
  width = 742,
  height = 1400,
  units = "px"
)

## First 5 PCs seems appropriate to me.
UMAP_Input <- df %>% 
  dplyr::select(starts_with("PC")) %>% 
  dplyr::select(1:5)

## Set seed for umap, if I understand correctly this is by default random and not the same seed as set before.
custom.config <- umap.defaults
custom.config$random_state <- 1337

## Create umap object from wide counts matrix

UMAP_Output <- umap(UMAP_Input,
  n_neighbors = 8
) # Highest value possible with 9 samples. Default is 15.

UMAP_mtx <- UMAP_Output[["layout"]]

colnames(UMAP_mtx) <- c("UMAP_1", "UMAP_2")

UMAP_df <- as.data.frame(UMAP_mtx) %>%
  rownames_to_column("User_ID") %>%
  inner_join(Sample_List,
    by = "User_ID"
  )

## Visualize umap
UMAP_Plot <- ggplot(
  data = UMAP_df,
  aes(
    x = UMAP_1,
    y = UMAP_2,
    color = condition
  )
) +
  scale_colour_manual(values = c("Non_Infected" = "#4DC36B", "Infected" = "#440C55")) +
  # geom_point() +
  geom_label(aes(label = Manuscript_ID),
             size = 2) +
  labs(
    title = "UMAP dimensionality reduction on PCA with PC1 through PC5"
  ) +
  theme_bw()
print(UMAP_Plot)

ggsave(
  filename = "./R_output_files/Figures/UMAP_Plot.eps",
  plot = UMAP_Plot,
  width = 1484,
  height = 700,
  units = "px"
)

################################## Differential Gene Expression (DGE)
## Extract DESeq2 results
res_df <- DESeq2::results(dds,
  alpha = 0.05,
  contrast = c("condition", "Infected", "Non_Infected")
)
summary(res_df)
## Independent Hypothesis Weighting http://bioconductor.org/packages/release/bioc/html/IHW.html
## Data frame from deseq results

deRes_res_df <- as.data.frame(res_df)

# IHW adjustment of pvalues
ihwRes_res_df <- ihw(pvalue ~ baseMean,
  data = deRes_res_df,
  alpha = 0.05
)

# Number of significant genes
paste(
  "The number of genes with adjusted p-value < 0.05 in Infected_vs_Non_Infected were",
  rejections(ihwRes_res_df),
  "with IHW and",
  sum(deRes_res_df$padj <= 0.05, na.rm = TRUE),
  "with BH"
)

# Dataframe with IHW adjusted p values, fold changes, counts and rlog
Combined_Results_DF <- as.data.frame(ihwRes_res_df,
  row.names = rownames(deRes_res_df)
) %>%
  rownames_to_column("GeneSymbol") %>%
  mutate(Comparison = "Infected_vs_Non_Infected") %>%
  inner_join(
    deRes_res_df %>%
      rownames_to_column("GeneSymbol"),
    by = c("GeneSymbol", "pvalue")
  ) %>%
  dplyr::select(GeneSymbol, adj_pvalue, pvalue, log2FoldChange, Comparison) %>%
  inner_join(Tidy_Counts_DetectedGenes_NoDuplicatedSymbol,
    by = "GeneSymbol",
    relationship = "many-to-many"
  ) %>%
  inner_join(
    as.data.frame(assay(rlog(dds, blind = TRUE))) %>%
      rownames_to_column("GeneSymbol") %>%
      pivot_longer(
        cols = starts_with("H9"),
        values_to = "Rlog_Counts",
        names_to = "User_ID"
      ),
    by = c("GeneSymbol", "User_ID")
  ) %>%
  inner_join(
    Sample_List %>%
      dplyr::select(User_ID, condition),
    by = "User_ID"
  )

# Differential gene expression volcano plot
Volcano_Plot_DF <- Combined_Results_DF %>%
  dplyr::filter(Gene_Detected == TRUE) %>%
  dplyr::select(GeneSymbol, adj_pvalue, log2FoldChange, Comparison) %>%
  distinct() %>%
  mutate(adj_pvalue = ifelse(is.na(adj_pvalue), 1, adj_pvalue)) %>%
  mutate(log2FoldChange = ifelse(is.na(log2FoldChange), 0, log2FoldChange)) %>%
  mutate(log2FoldChange = ifelse(log2FoldChange < -8, -8, log2FoldChange)) %>% # A few non-significant genes have extreme fold changes, this is for better visualisation
  mutate(log2FoldChange = ifelse(log2FoldChange > 8, 8, log2FoldChange)) %>% # A few non-significant genes have extreme fold changes, this is for better visualisation
  mutate(Regulation = factor(case_when(
    log2FoldChange > 0 & adj_pvalue < 0.05 ~ "up",
    log2FoldChange < -0 & adj_pvalue < 0.05 ~ "down",
    TRUE ~ "neither"
  )))

cols <- c("up" = "#ffad73", "down" = "#26b3ff", "neither" = "grey")
sizes <- c("up" = 3, "down" = 3, "neither" = 1)
alphas <- c("up" = 1, "down" = 1, "neither" = 0.5)

Volcano_Plot <- ggplot(
  data = Volcano_Plot_DF,
  aes(
    x = log2FoldChange,
    y = -log10(adj_pvalue),
    size = Regulation,
    colour = Regulation,
    alpha = Regulation
  )
) +
  geom_point() +
  scale_colour_manual(values = cols) +
  scale_size_manual(values = sizes) +
  scale_alpha_manual(values = alphas) +
  geom_label(
    data = Volcano_Plot_DF %>%
      dplyr::filter(adj_pvalue < 0.05) %>%
      dplyr::filter(log2FoldChange > 2 | log2FoldChange < -2),
    aes(label = GeneSymbol),
    label.size = 1
  ) +
  geom_hline(
    yintercept = -log10(0.05),
    linetype = "dashed"
  ) +
  geom_vline(
    xintercept = c(-2, 2),
    linetype = "dashed"
  ) +
  scale_x_continuous(
    breaks = c(seq(-8, 8, 1)),
    limits = c(-8, 8)
  ) +
  scale_y_continuous(
    breaks = c(seq(0, 12, 1), -log10(0.05)),
    limits = c(0, 12),
    labels = as.character(c(seq(0, 12, 1), "adjusted P=0.05"))
  ) +
  labs(
    title = "Gene expression changes in infected vs non-infected samples",
    x = "log2(fold change)",
    y = "-log10(adjusted P-value)"
  ) +
  theme_bw() +
  theme(
    panel.border = element_rect(colour = "black", fill = NA),
    panel.grid.minor = element_blank(),
    panel.grid.major = element_blank()
  )
print(Volcano_Plot)

ggsave(
  filename = "./R_output_files/Figures/Volcano_Plot_Differential_Gene_Expression_Infection.svg",
  plot = Volcano_Plot,
  width = 2226,
  height = 1205,
  units = "px"
)

ggsave(
  filename = "./R_output_files/Figures/Volcano_Plot_Differential_Gene_Expression_Infection.pdf",
  plot = Volcano_Plot,
  width = 2226,
  height = 1205,
  units = "px"
)

## Saving data for all genes
genes_for_annotation <- Combined_Results_DF %>%
  dplyr::select(GENEID, GeneSymbol, Gene_Detected) %>%
  distinct() %>%
  replace_na(list(log2FoldChange = 0, adj_pvalue = 1, pvalue = 1)) # NA values might disrupt downstream analysis

all_detected_genes <- Combined_Results_DF %>%
  dplyr::filter(Gene_Detected == TRUE) %>%
  dplyr::select(GENEID, GeneSymbol, log2FoldChange, adj_pvalue, pvalue, Comparison) %>%
  distinct() %>%
  replace_na(list(log2FoldChange = 0, adj_pvalue = 1, pvalue = 1)) # NA values might disrupt downstream analysis

sig_genes <- all_detected_genes %>%
  dplyr::filter(adj_pvalue < 0.05)
################################## Export
write_tsv(all_detected_genes, "./R_output_files/Tables/All_Detected_Genes.tsv")

write_tsv(
  sig_genes,
  "./R_output_files/Tables/Significant_Genes.tsv"
)

write_xlsx(
  x = sig_genes,
  path = "./R_output_files/Tables/Significant_Genes.xlsx"
)

# top 10 lowest padj
sig_genes %>%
  slice_min(
    order_by = adj_pvalue,
    n = 10
  )

# Significant genes with log2 fold change > 2
sig_genes %>%
  dplyr::filter(log2FoldChange > 2)

# Significant genes with log2 fold change < -2
sig_genes %>%
  dplyr::filter(log2FoldChange < -2)

# Significant genes with absolute log 2 fold change > 4
top_infection_genes_fc4 <- sig_genes %>%
  dplyr::filter(log2FoldChange > 4 | log2FoldChange < -4) %>%
  arrange(log2FoldChange) %>%
  dplyr::select(GeneSymbol, log2FoldChange, adj_pvalue)

# Exporting significant genes with absolute log 2 fold change
write_xlsx(top_infection_genes_fc4, "./R_output_files/Tables/top_infection_genes_fc4.xlsx")

# Significant genes with absolute log 2 fold change > 3
top_infection_genes_fc3 <- sig_genes %>%
  dplyr::filter(log2FoldChange > 3 | log2FoldChange < -3) %>%
  arrange(log2FoldChange) %>%
  dplyr::select(GeneSymbol, log2FoldChange, adj_pvalue)

# Exporting significant genes with absolute log 2 fold change
write_xlsx(top_infection_genes_fc3, "./R_output_files/Tables/top_infection_genes_fc3.xlsx")

# Significant genes with absolute log 2 fold change > 2
top_infection_genes_fc2 <- sig_genes %>%
  dplyr::filter(log2FoldChange > 2 | log2FoldChange < -2) %>%
  arrange(log2FoldChange) %>%
  dplyr::select(GeneSymbol, log2FoldChange, adj_pvalue)

# Exporting significant genes with absolute log 2 fold change
write_xlsx(top_infection_genes_fc2, "./R_output_files/Tables/top_infection_genes_fc2.xlsx")


################################## Calculate TPM (transcripts per kilobase million) for better visualization of within-sample gene expression
ensdb <- EnsDb.Mmusculus.v79

## List all supported keytypes.
keytypes(ensdb)
## List all supported columns for the select and mapIds methods.
columns(ensdb)
## List /real/ database column names.
listColumns(ensdb)
## Retrieve all keys corresponding to keytype
ENTREZID <- keys(ensdb, keytype = "ENTREZID")
length(ENTREZID)
head(ENTREZID)
GENEID <- keys(ensdb, keytype = "GENEID")
length(GENEID)
head(GENEID)
GENENAME <- keys(ensdb, keytype = "GENENAME")
length(GENENAME)
head(GENENAME)
SYMBOL <- keys(ensdb, keytype = "SYMBOL")
length(SYMBOL)
head(SYMBOL)
TXID <- keys(ensdb, keytype = "TXID")
length(TXID)
head(TXID)

## Genename and symbol look identical, are they?
length(SYMBOL) == length(GENENAME)
SYMBOL[!SYMBOL == GENENAME]
## They are identical

## Lets get a mapping table between ensdb and our data

ensdb_gene_counts <- enframe(
  x = mapIds(ensdb, keys = SYMBOL, column = "GENEID", keytype = "SYMBOL"),
  name = "GeneSymbol",
  value = "ensembldb_ID"
) %>%
  inner_join(Tidy_Counts_DetectedGenes_NoDuplicatedSymbol,
    by = "GeneSymbol"
  ) %>%
  inner_join(
    enframe(
      x = mapIds(ensdb, keys = SYMBOL, column = "TXID", keytype = "SYMBOL"),
      name = "GeneSymbol",
      value = "ensembldb_TXID"
    ),
    by = "GeneSymbol"
  )

## Now get transcript lengths and calculate RPK (reads per kilobase)
ensdb_gene_counts_transcripts <- enframe(
  lengthOf(
    x = ensdb,
    of = "tx"
  ),
  name = "ensembldb_TXID",
  value = "Transcript_Length"
) %>%
  inner_join(ensdb_gene_counts,
    by = "ensembldb_TXID"
  ) %>%
  mutate(Transcript_Kbp = Transcript_Length / 1000) %>%
  mutate(RPK = Counts / Transcript_Kbp)

## Calculate scaling factor for TPM (transcripts per million), calculate TPM, and add more annotation
TPM_df <- ensdb_gene_counts_transcripts %>%
  group_by(User_ID) %>%
  dplyr::summarise(TPM_Scaling_Factor = sum(RPK) / 1e6) %>%
  inner_join(ensdb_gene_counts_transcripts,
    by = "User_ID"
  ) %>%
  mutate(TPM = RPK / TPM_Scaling_Factor) %>%
  inner_join(Sample_List,
    by = "User_ID"
  )

# exporting TPM_df intermediate file
write_tsv(
  x = TPM_df,
  file = "./R_intermediate_files/TPM_df.tsv"
)

########################################## Gene markers for cell types to visualize what cells might be present
## PanglaoDB marker genes download
PanglaoDB_Marker_Genes <- read_tsv(
  file = "./R_input_files/PanglaoDB_markers_27_Mar_2020.tsv",
  col_types = c("ccccdcciccdddd")
) %>%
  dplyr::rename("GeneSymbol" = `official gene symbol`) %>%
  dplyr::rename("Cell_Type" = `cell type`) %>%
  dplyr::mutate(GeneSymbol = str_to_title(GeneSymbol))

## Organ filter for cell types that could feasibly exist in the tissue
Organ_Filter <- c(
  "GI tract",
  "Immune system"
)

## Manually adding genes based on Bockerstett et al 2020 https://gut.bmj.com/content/69/6/1027.long
Bockerstett_Gene_List <- read_tsv(
  file = "./R_input_files/Bockerstett et all 2020 Gut supplemental material marker gene list.tsv",
  col_types = c("cic")
) %>%
  dplyr::select(GeneSymbol, Cell_Type) %>%
  mutate("organ" = "GI tract") %>%
  distinct()

## Filter for genes presented as marker genes in several cell types in the cell types selected
Gene_Filter <- PanglaoDB_Marker_Genes %>%
  dplyr::filter(organ %in% Organ_Filter) %>%
  full_join(Bockerstett_Gene_List,
    by = c("GeneSymbol", "Cell_Type", "organ")
  ) %>%
  dplyr::count(GeneSymbol) %>%
  dplyr::filter(n != 1) %>%
  dplyr::select(GeneSymbol) %>%
  deframe()

## Cell type filter for cell types that could feasibly exist in the tissue
Cell_Types_Filter <- c(
  "B cells",
  "Basophils",
  "Crypt cells",
  "Dendritic cells",
  "Endothelial cells",
  "Enteroendocrine cells",
  "Eosinophils",
  "Fibroblasts",
  "Foveolar cells",
  "Gastric chief cells",
  "Macrophages",
  "Mast cells",
  "Metaplastic cells",
  "Monocytes",
  "Mucous neck cells",
  "Natural killer T cells",
  "Enterochromaffin cells",
  # "Neuroendocrine cells", #These marker genes overlap with other genes so this cell type wouldn't show up after downstream filters
  "Neutrophils",
  "Stromal cells",
  "Parietal cells",
  "Pericytes",
  "Plasma cells",
  "Plasmacytoid dendritic cells",
  "Proliferating Tff2+ cells", # A version of metaplastic cells
  "T cells",
  "Tuft cells"
)

## Filter marker genes which show up in multiple cell types in our analysis in order to increase specificity
Gene_Filter <- PanglaoDB_Marker_Genes %>%
  full_join(Bockerstett_Gene_List,
    by = c("GeneSymbol", "Cell_Type", "organ")
  ) %>%
  dplyr::filter(GeneSymbol != "Muc6") %>% # Manual removal from gene filter. Will instead be removed from metaplastic cells to avoid overlap
  dplyr::filter(Cell_Type %in% Cell_Types_Filter) %>%
  dplyr::count(GeneSymbol) %>%
  dplyr::filter(n != 1) %>%
  dplyr::select(GeneSymbol) %>%
  deframe()

## DF of marker genes
TPM_ggplot_df <- PanglaoDB_Marker_Genes %>%
  dplyr::filter(`canonical marker` == 1) %>% # One way to get as specific gene markers as possible
  dplyr::filter(!specificity_mouse > 0) %>% # Another way to make gene markers more specific
  dplyr::filter(species == "Mm Hs" | species == "Mm") %>% # We only want marker genes relevant for mice
  full_join(Bockerstett_Gene_List,
    by = c("GeneSymbol", "Cell_Type", "organ")
  ) %>%
  dplyr::filter(!(GeneSymbol == "Muc6" & Cell_Type == "Metaplastic cells")) %>% # Filtering out Muc6 for metaplastic cells to avoid overlap with mucous neck cells
  dplyr::filter(!GeneSymbol %in% Gene_Filter) %>%
  dplyr::filter(Cell_Type %in% Cell_Types_Filter) %>%
  inner_join(TPM_df,
    by = "GeneSymbol",
    relationship = "many-to-many"
  ) %>%
  dplyr::select(Cell_Type, GeneSymbol, TPM, names, condition, organ, `canonical marker`, specificity_mouse) %>%
  distinct() %>% 
  mutate(TPM_Log10 = log10(TPM)) %>% 
  mutate(TPM_Log2 = log2(TPM)) %>% 
  group_by(Cell_Type) %>% 
  mutate(Mean_Cell_Type_TPM = mean(TPM)) %>% 
  ungroup()

## How many marker genes are there for each cell type?
TPM_ggplot_df %>%
  dplyr::select(Cell_Type, GeneSymbol) %>%
  distinct() %>%
  plyr::count("Cell_Type")

## Save table of marker gene overview
write_xlsx(
  x = TPM_ggplot_df %>%
    dplyr::select(Cell_Type, GeneSymbol) %>%
    dplyr::distinct() %>%
    plyr::count("Cell_Type"),
  path = "./R_output_files/Tables/Cell_Type_Markers_Summary.xlsx"
)

## Save table of marker gene details
write_xlsx(
  x = TPM_ggplot_df %>%
    dplyr::select(Cell_Type, GeneSymbol) %>%
    dplyr::distinct() %>% 
    dplyr::arrange(Cell_Type),
  path = "./R_output_files/Tables/Cell_Type_Markers_Details.xlsx"
)

## Plot marker genes
Marker_Genes_Plot <- ggplot(
  TPM_ggplot_df,
  aes(
    x = Cell_Type,
    y = TPM,
    colour = condition
  )
) +
  geom_point() +
  theme_bw()
Marker_Genes_Plot

## Genes that stick out are TFF1, MUC5AC, ATP4A, ATP4B, and Tspan8.
## Add manual annotation for these genes

TPM_ggplot_df_annotated <- TPM_ggplot_df %>%
  mutate(High_TPM_Genes = if_else(GeneSymbol %in% c("Muc5ac", "Tff1", "Atp4a", "Atp4b", "Tspan8"), GeneSymbol, "None")) # These are the top 5 genes, and they have an expression greater than 1000TPM

Marker_Genes_Plot <- ggplot(
  TPM_ggplot_df_annotated,
  aes(
    x = Cell_Type,
    y = TPM_Log10,  # Extreme differences in TPM, better to visualise on log scale
    shape = condition,
    color = High_TPM_Genes
  )
) +
  geom_jitter(width = 0.25) +
  scale_y_continuous(breaks = seq(0, max(TPM_ggplot_df$TPM_Log10), by = 1)) +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  geom_hline(yintercept = 3)
Marker_Genes_Plot

## Save the plot
ggsave(
  filename = "./R_output_files/Figures/Marker_Genes_Plot.eps",
  plot = Marker_Genes_Plot,
  width = 2230,
  height = 1300,
  units = "px"
)

## Are any of the marker genes among genes with significant differential gene expression?
Significant_Marker_Genes <- sig_genes %>%
  # mutate(GeneSymbol = str_to_upper(GeneSymbol)) %>%
  dplyr::filter(GeneSymbol %in% TPM_ggplot_df_annotated$GeneSymbol)
Significant_Marker_Genes

## Find which cell types these belong to
Significant_Cell_Types <- TPM_ggplot_df_annotated %>%
  dplyr::filter(GeneSymbol %in% Significant_Marker_Genes$GeneSymbol) %>%
  dplyr::select(Cell_Type, GeneSymbol) %>%
  distinct()
Significant_Cell_Types

## How many marker genes do these cell types have in total?
TPM_ggplot_df_annotated %>%
  dplyr::filter(Cell_Type %in% Significant_Cell_Types$Cell_Type) %>%
  dplyr::select(Cell_Type, GeneSymbol) %>%
  distinct() %>%
  plyr::count("Cell_Type")

########################################## Investigate which genes are the top-most expressed in the non-infected mice
Normal_High_Abundance_Genes <- TPM_df %>% 
  dplyr::filter(condition == "Non_Infected") %>% 
  group_by(GeneSymbol) %>% 
  mutate(Mean_GeneSymbol_TPM = mean(TPM)) %>% 
  ungroup() %>% 
  dplyr::filter(TPM > 4000) # The lowest TPM of Muc5ac was 4689. The highest TPM of non-foveolar marker genes was 1479 (Atp4a). 4000 is below the Muc5ac, yet much higher than the highest non-foveolar gene. This should strike a balance between specificity and sensitivity fo the description of a normal mouse gastric corpus foveolar cell.

# How many genes would this be?
Normal_High_Abundance_Genes %>% 
  dplyr::select(GeneSymbol, Mean_GeneSymbol_TPM) %>% 
  distinct()
# 29 genes

# What about with a cutoff using the lowest Muc5ac expression?
Normal_High_Abundance_Genes <- TPM_df %>% 
  dplyr::filter(condition == "Non_Infected") %>% 
  dplyr::filter(TPM >= min(TPM_df %>% 
                  dplyr::filter(GeneSymbol == "Muc5ac") %>% 
                  dplyr::select(TPM) %>% 
                  deframe())) %>% 
  dplyr::select(GeneSymbol) %>% 
  distinct() %>% 
  inner_join(TPM_df, by = "GeneSymbol") %>% 
  dplyr::filter(condition == "Non_Infected") %>% 
  mutate(TPM_Log10 = log10(TPM)) %>% 
  group_by(GeneSymbol) %>% 
  mutate(Mean_GeneSymbol_TPM = mean(TPM)) %>% 
  mutate(Mean_GeneSymbol_TPM_Log10 = mean(TPM_Log10)) %>% 
  ungroup() %>% 
  inner_join(all_detected_genes, by = c("GeneSymbol", "GENEID"))

# How many genes would this be?
Normal_High_Abundance_Genes %>% 
  dplyr::select(GeneSymbol, Mean_GeneSymbol_TPM) %>% 
  distinct()
# 26 genes

# What about with a cutoff using the lowest Muc5ac expression, but with the mean of the gene has to be higher instead of just the highest TPM of all samples has to be higher?
Normal_High_Abundance_Genes <- TPM_df %>% 
  dplyr::filter(condition == "Non_Infected") %>% 
  mutate(TPM_Log10 = log10(TPM)) %>% 
  group_by(GeneSymbol) %>% 
  mutate(Mean_GeneSymbol_TPM = mean(TPM)) %>% 
  mutate(Mean_GeneSymbol_TPM_Log10 = mean(TPM_Log10)) %>% 
  ungroup() %>% 
  dplyr::filter(Mean_GeneSymbol_TPM >= min(TPM_df %>% 
                             dplyr::filter(GeneSymbol == "Muc5ac") %>% 
                             dplyr::select(TPM) %>% 
                             deframe())) %>% 
  distinct() %>% 
  inner_join(all_detected_genes, by = c("GeneSymbol", "GENEID"))

# How many genes would this be?
Normal_High_Abundance_Genes %>% 
  dplyr::select(GeneSymbol, Mean_GeneSymbol_TPM) %>% 
  distinct()
# 18 genes

# 18 genes is feasible for manual review. The cutoff is rather strict, but makes biological sense. And it allows the possibility of inclusion of genes with lower TPM than Muc5ac.

# Visualize
Normal_High_Abundance_Genes_Plot <- ggplot(
  Normal_High_Abundance_Genes,
  aes(
    x = GeneSymbol,
    y = TPM_Log10,  # Extreme differences in TPM, better to visualise on log scale
  )
) +
  geom_jitter(width = 0.25) +
  scale_y_continuous(breaks = seq(0, max(Normal_High_Abundance_Genes$TPM_Log10), by = 0.1)) +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
Normal_High_Abundance_Genes_Plot

# Compress data frame with TPM and Gene Symbol
Normal_High_Abundance_Genes <- Normal_High_Abundance_Genes %>% 
  dplyr::select(GeneSymbol, Mean_GeneSymbol_TPM) %>% 
  distinct() %>% 
  arrange(desc(Mean_GeneSymbol_TPM)) %>% 
  mutate("High_Abundance_TPM" = TRUE)

# Add info if marker gene according to PanglaoDb and/or Bockerstett, and get TPM for all marker genes no matter source
# The potential marker genes according to our data that has high TPM is marked as high abundance tpm
Mean_GeneSymbol_TPM_Non_Infected <- TPM_df %>% 
  dplyr::filter(condition == "Non_Infected") %>% 
  group_by(GeneSymbol) %>% 
  mutate(Mean_GeneSymbol_TPM = mean(TPM)) %>% 
  ungroup() %>% 
  dplyr::select(GeneSymbol, Mean_GeneSymbol_TPM) %>% 
  distinct()

Normal_High_Abundance_Genes_PanglaoDB <- PanglaoDB_Marker_Genes %>% 
  dplyr::filter(Cell_Type == "Foveolar cells") %>% 
  dplyr::filter(str_detect(species, "Mm")) %>% 
  mutate("PanglaoDB_Canonical_Marker" = ifelse(is.na(`canonical marker`), FALSE, TRUE)) %>% 
  dplyr::select(GeneSymbol, PanglaoDB_Canonical_Marker) %>% 
  dplyr::mutate(PanglaoDB_Foveolar_Cell_Marker = TRUE) %>% 
  inner_join(Mean_GeneSymbol_TPM_Non_Infected, by = "GeneSymbol")

Normal_High_Abundance_Genes_Bockerstett <- Bockerstett_Gene_List %>% 
  dplyr::filter(Cell_Type == "Foveolar cells") %>% 
  dplyr::select(GeneSymbol) %>% 
  dplyr::mutate(Bockerstett_Foveolar_Cell_Marker = TRUE) %>% 
  inner_join(Mean_GeneSymbol_TPM_Non_Infected, by = "GeneSymbol")

Normal_High_Abundance_Genes_And_Marker_Genes <- Normal_High_Abundance_Genes %>% 
  dplyr::full_join(Normal_High_Abundance_Genes_PanglaoDB, by = c("GeneSymbol", "Mean_GeneSymbol_TPM")) %>% 
  dplyr::full_join(Normal_High_Abundance_Genes_Bockerstett, by = c("GeneSymbol", "Mean_GeneSymbol_TPM")) %>% 
  dplyr::mutate(across(everything(), ~ replace_na(.x, replace = FALSE))) %>% 
  arrange(desc(Mean_GeneSymbol_TPM)) %>% 
  mutate("Mean_GeneSymbol_TPM" = round(Mean_GeneSymbol_TPM))

# Export table
write_xlsx(x = Normal_High_Abundance_Genes_And_Marker_Genes,
           path = "./R_output_files/Tables/Normal_High_Abundance_Genes_And_Marker_Genes.xlsx")


########################################## Comparison of TPM of significant genes
## Complementary analysis of a more absolute meassure rather relative gene expression
TPM_sig_genes <- TPM_df %>%
  dplyr::select(GeneSymbol, User_ID, TPM, condition) %>%
  dplyr::group_by(GeneSymbol, condition) %>%
  dplyr::mutate(Mean_TPM = mean(TPM)) %>%
  dplyr::mutate(condition = factor(condition, levels = c("Non_Infected", "Infected"))) %>%
  dplyr::ungroup() %>%
  dplyr::inner_join(sig_genes, by = "GeneSymbol") %>%
  dplyr::filter(adj_pvalue < 0.05)

TPM_sig_genes_plot <- ggplot(TPM_sig_genes, aes(x = condition, y = Mean_TPM, group = GeneSymbol)) +
  geom_line() +
  geom_label(aes(label = GeneSymbol)) +
  theme_bw()
print(TPM_sig_genes_plot)

ggsave(
  filename = "./R_output_files/Figures/TPM_sig_genes_plot.eps",
  plot = TPM_sig_genes_plot,
  width = 1100,
  height = 1305,
  units = "px"
)

# Mean TPM df for downstream
Mean_condition_TPM_sig_genes <- TPM_sig_genes %>%
  dplyr::select(GeneSymbol, Mean_TPM, condition) %>%
  distinct() %>%
  pivot_wider(names_from = condition, values_from = Mean_TPM)

# Exporting mean TPM intermediary file
write_tsv(
  x = Mean_condition_TPM_sig_genes,
  file = "./R_intermediate_files/Mean_condition_TPM_sig_genes.tsv"
)

##################################################### Next comes several sections for annotating the genes with gene sets. It is very messy, to whoever is re-running this script I'm so sorry about this
##################################################### Annotate GO gene sets with bioMart, might be better than annotationdbi
## Annotate with GO.db and Biomart and compare
## First Biomart
## The rnaseq pipeline used ENSEMBL release 81
listEnsemblArchives()
## BiomaRt doesn't have release 81. I will use the latest realease which of writing is 109. It should have better annotation with for example GO terms.
listEnsembl(version = 109)
ensembl <- useEnsembl(
  biomart = "genes",
  version = 109
)
listDatasets(ensembl)
searchDatasets(mart = ensembl, pattern = "mus")
ensembl <- useEnsembl(
  biomart = "genes",
  dataset = "mmusculus_gene_ensembl",
  host = "www.ensembl.org",
  mirror = "www"
)
head(listFilters(ensembl))
Attributes <- listAttributes(ensembl)
Legend_bioMart <- getBM(
  mart = ensembl,
  attributes = c(
    "ensembl_gene_id",
    "external_gene_name"
  )
)

# Not all ensembl IDs and gene symbols will match perfectly.
# Creating an annotation dataframe that is as complete as possible by taking into account mis-match
# First, the genes where both identifiers match
Matching_ENSEMBL_Symbol <- genes_for_annotation %>%
  mutate("ensembl_gene_id" = GENEID) %>%
  mutate("external_gene_name" = GeneSymbol) %>%
  inner_join(Legend_bioMart,
    by = c("ensembl_gene_id", "external_gene_name")
  ) %>%
  mutate(biomaRt_compatability = "ENSEMBL_and_symbol_matching") %>%
  relocate(c(
    GENEID,
    ensembl_gene_id,
    GeneSymbol,
    external_gene_name,
    biomaRt_compatability
  )) %>%
  distinct()

# Where only ENSEMBL match
Matching_ENSEMBL_Only <- genes_for_annotation %>%
  dplyr::filter(!GENEID %in% Matching_ENSEMBL_Symbol$GENEID) %>%
  mutate("ensembl_gene_id" = GENEID) %>%
  inner_join(Legend_bioMart,
    by = c("ensembl_gene_id")
  ) %>%
  mutate(biomaRt_compatability = "ENSEMBL_only_matching") %>%
  relocate(c(
    GENEID,
    ensembl_gene_id,
    GeneSymbol,
    external_gene_name,
    biomaRt_compatability
  )) %>%
  distinct()

# Where only symbol match
Matching_Symbol_Only <- genes_for_annotation %>%
  dplyr::filter(!GENEID %in% Matching_ENSEMBL_Symbol$GENEID) %>%
  dplyr::filter(!GENEID %in% Matching_ENSEMBL_Only$GENEID) %>%
  mutate("external_gene_name" = GeneSymbol) %>%
  inner_join(Legend_bioMart,
    by = c("external_gene_name"),
    multiple = "first"
  ) %>% # How I handled multiple matches
  mutate(biomaRt_compatability = "Symbol_only_matching") %>%
  relocate(c(
    GENEID,
    ensembl_gene_id,
    GeneSymbol,
    external_gene_name,
    biomaRt_compatability
  )) %>%
  distinct()

# Where neither ENSEMBL nor symbol matched
No_Matching <- genes_for_annotation %>%
  dplyr::filter(!GENEID %in% Matching_ENSEMBL_Symbol$GENEID) %>%
  dplyr::filter(!GENEID %in% Matching_ENSEMBL_Only$GENEID) %>%
  dplyr::filter(!GENEID %in% Matching_Symbol_Only$GENEID) %>%
  mutate("ensembl_gene_id" = GENEID) %>%
  mutate("external_gene_name" = GeneSymbol) %>%
  mutate(biomaRt_compatability = "No_matching") %>%
  relocate(c(
    GENEID,
    ensembl_gene_id,
    GeneSymbol,
    external_gene_name,
    biomaRt_compatability
  )) %>%
  distinct()

GeneSymbol_biomaRt_Annotated_ENSEMBL_Symbol <- rbind(
  Matching_ENSEMBL_Symbol,
  Matching_ENSEMBL_Only,
  Matching_Symbol_Only,
  No_Matching
) %>%
  arrange(GENEID)

# If need to split up:
Entrez <- getBM(
  mart = ensembl,
  attributes = c(
    "ensembl_gene_id",
    "entrezgene_id",
    "entrezgene_accession",
    "entrezgene_description"
  )
)

GO1 <- getBM(
  mart = ensembl,
  attributes = c(
    "ensembl_gene_id",
    "go_id"
  )
)

GO2 <- getBM(
  mart = ensembl,
  attributes = c(
    "go_id", # GO term accession
    "name_1006", # GO term name
    "definition_1006", # GO term definition
    # "go_linkage_type", #GO term evidence code
    "namespace_1003" # GO domain
  )
)

GO <- GO1 %>%
  inner_join(GO2,
    by = "go_id",
    multiple = "all"
  ) %>%
  dplyr::filter(go_id != "") %>%
  dplyr::filter(namespace_1003 != "")


GO_Entrez <- Entrez %>%
  mutate(entrezgene_id = as.character(entrezgene_id)) %>%
  full_join(GO,
    by = "ensembl_gene_id",
    relationship = "many-to-many"
  ) %>%
  dplyr::mutate(entrezgene_id = na_if(entrezgene_id, "")) %>%
  dplyr::mutate(entrezgene_accession = na_if(entrezgene_accession, "")) %>%
  dplyr::mutate(entrezgene_description = na_if(entrezgene_description, ""))

# Annotation with ENSEMBL, symbols, Entrez and GO all in one
BiomaRt_ENSEMBL_Symbol_Entrez_GO_Annotation <- GeneSymbol_biomaRt_Annotated_ENSEMBL_Symbol %>%
  full_join(GO_Entrez,
    by = "ensembl_gene_id",
    relationship = "many-to-many"
  )

# nrow(BiomaRt_ENSEMBL_Symbol_Entrez_GO_Annotation %>% dplyr::select(GENEID) %>% distinct()) == nrow(GeneSymbol_biomaRt_Annotated_ENSEMBL_Symbol %>% dplyr::select(GENEID) %>% distinct())
Genes_Without_BiomaRt_GO_Annotation <- BiomaRt_ENSEMBL_Symbol_Entrez_GO_Annotation[which(is.na(BiomaRt_ENSEMBL_Symbol_Entrez_GO_Annotation$go_id)), ]
# 19k genes lack GO-annotation in the ensembl database
sig_genes %>%
  dplyr::filter(GENEID %in% Genes_Without_BiomaRt_GO_Annotation$GENEID)
# significant genes lacking GO annotation.
# Browsing the ENSEMBL genome browser, they actually do lack GO annotation in the ENSEMBL database.
# Meanwhile, some of them are part of reactome pathways

Genes_Without_Entrez <- BiomaRt_ENSEMBL_Symbol_Entrez_GO_Annotation %>%
  dplyr::filter(is.na(entrezgene_id)) %>%
  dplyr::select(GENEID) %>%
  distinct()

sig_genes %>%
  dplyr::filter(GENEID %in% Genes_Without_Entrez$GENEID)
# signifcant genes lacking entrez id.

##################################################### Annotate with annotationdbi, ENSEMBLs database which biomart connects to is missing some GO terms
keytypes(org.Mm.eg.db)
head(keys(x = org.Mm.eg.db, keytype = "ENTREZID"))
columns(org.Mm.eg.db)
Annotationdbi_ENSEMBL <- keys(org.Mm.eg.db, keytype = "ENSEMBL")
# How many ENSEMBL ids match?
nrow(genes_for_annotation %>%
  dplyr::filter(GENEID %in% Annotationdbi_ENSEMBL))
# How many symbols match?
Annotationdbi_SYMBOL <- keys(org.Mm.eg.db, keytype = "SYMBOL")
nrow(genes_for_annotation %>%
  dplyr::filter(GeneSymbol %in% Annotationdbi_SYMBOL))
# Annotationdbi works on an entrez framework.
# Compare entrez conversion in annotationdbi with biomart
Annotationdbi_ENTREZ <- keys(x = org.Mm.eg.db, keytype = "ENTREZID")

Annotationdbi_ENTREZ_ENSEMBL_SYMBOL <- AnnotationDbi::select(
  x = org.Mm.eg.db,
  keys = Annotationdbi_ENTREZ,
  columns = c("ENSEMBL", "SYMBOL")
)
nrow(genes_for_annotation %>%
  dplyr::filter(GeneSymbol %in% Annotationdbi_ENTREZ_ENSEMBL_SYMBOL$SYMBOL))
# matches by symbol in annotationdbi
nrow(genes_for_annotation %>%
  dplyr::filter(GENEID %in% Annotationdbi_ENTREZ_ENSEMBL_SYMBOL$ENSEMBL))
# matches by ensembl in annotationdbi
nrow(
  BiomaRt_ENSEMBL_Symbol_Entrez_GO_Annotation %>%
    dplyr::filter(!is.na(entrezgene_id)) %>%
    dplyr::select(GENEID, GeneSymbol, entrezgene_id) %>%
    distinct()
)
# matches with biomart

# Match by ensembl IDs and symbols
Annotationdbi_Matching_ENSEMBL_Symbol <- genes_for_annotation %>%
  mutate("ENSEMBL" = GENEID) %>%
  mutate("SYMBOL" = GeneSymbol) %>%
  inner_join(Annotationdbi_ENTREZ_ENSEMBL_SYMBOL,
    by = c("ENSEMBL", "SYMBOL")
  ) %>%
  mutate(Annotationdbi_compatability = "ENSEMBL_and_SYMBOL_matching") %>%
  relocate(c(
    GENEID,
    ENSEMBL,
    GeneSymbol,
    SYMBOL,
    ENTREZID,
    Annotationdbi_compatability
  )) %>%
  distinct()

# Where only ENSEMBL match
Annotationdbi_Matching_ENSEMBL_Only <- genes_for_annotation %>%
  dplyr::filter(!GENEID %in% Annotationdbi_Matching_ENSEMBL_Symbol$GENEID) %>%
  mutate("ENSEMBL" = GENEID) %>%
  inner_join(Annotationdbi_ENTREZ_ENSEMBL_SYMBOL,
    by = c("ENSEMBL"),
    relationship = "many-to-many"
  ) %>%
  mutate(Annotationdbi_compatability = "ENSEMBL_matching") %>%
  relocate(c(
    GENEID,
    ENSEMBL,
    GeneSymbol,
    SYMBOL,
    ENTREZID,
    Annotationdbi_compatability
  )) %>%
  distinct()


# Where only symbol match
Annotationdbi_Matching_SYMBOL_Only <- genes_for_annotation %>%
  dplyr::filter(!GENEID %in% Annotationdbi_Matching_ENSEMBL_Symbol$GENEID) %>%
  dplyr::filter(!GENEID %in% Annotationdbi_Matching_ENSEMBL_Only$GENEID) %>%
  mutate("SYMBOL" = GeneSymbol) %>%
  inner_join(Annotationdbi_ENTREZ_ENSEMBL_SYMBOL,
    by = c("SYMBOL"),
    relationship = "many-to-many"
  ) %>%
  mutate(Annotationdbi_compatability = "SYMBOL_matching") %>%
  relocate(c(
    GENEID,
    ENSEMBL,
    GeneSymbol,
    SYMBOL,
    ENTREZID,
    Annotationdbi_compatability
  )) %>%
  distinct()

# Where neither ENSEMBL nor symbol matched
Annotationdbi_Matching_Nothing_Only <- genes_for_annotation %>%
  dplyr::filter(!GENEID %in% Annotationdbi_Matching_ENSEMBL_Symbol$GENEID) %>%
  dplyr::filter(!GENEID %in% Annotationdbi_Matching_ENSEMBL_Only$GENEID) %>%
  dplyr::filter(!GENEID %in% Annotationdbi_Matching_SYMBOL_Only$GENEID) %>%
  mutate("ENSEMBL" = NA) %>%
  mutate("SYMBOL" = NA) %>%
  mutate("ENTREZID" = NA) %>%
  mutate(Annotationdbi_compatability = "No_matching") %>%
  relocate(c(
    GENEID,
    ENSEMBL,
    GeneSymbol,
    SYMBOL,
    ENTREZID,
    Annotationdbi_compatability
  )) %>%
  distinct()

# Bind dataframes together of all possible matches/non-matches
Annotationdbi_Annotated_ENTREZ_ENSEMBL_SYMBOL <- rbind(
  Annotationdbi_Matching_ENSEMBL_Symbol,
  Annotationdbi_Matching_ENSEMBL_Only,
  Annotationdbi_Matching_SYMBOL_Only,
  Annotationdbi_Matching_Nothing_Only
) %>%
  arrange(GENEID)

## Now GO terms with annotationdbi
keytypes(org.Mm.eg.db)
# help("GO")
head(keys(x = org.Mm.eg.db, keytype = "GO"))
head(keys(x = org.Mm.eg.db, keytype = "GOALL"))
columns(org.Mm.eg.db)
# Annotationdbi_GO <- keys(org.Mm.eg.db, keytype = "GO")
Annotationdbi_GOALL <- keys(org.Mm.eg.db, keytype = "GOALL")
# GO_ENTREZID <- AnnotationDbi::select(x = org.Mm.eg.db,
#                      keys = Annotationdbi_GO,
#                      column = c("ENTREZID"),
#                      keytype = "GO") %>%
# dplyr::select(-EVIDENCE) %>%
#  distinct()
GOALL_ENTREZID <- AnnotationDbi::select(
  x = org.Mm.eg.db,
  keys = Annotationdbi_GOALL,
  column = c("ENTREZID"),
  keytype = "GOALL"
) %>%
  dplyr::select(-EVIDENCEALL) %>%
  distinct()

# org.Mm.eg.db does not seem to have GO names, have to use GO.db for this
keytypes(GO.db)
# help("TERM")
GO_Annotation <- AnnotationDbi::select(
  x = GO.db,
  keys = Annotationdbi_GOALL,
  column = c("TERM", "DEFINITION"),
  keytype = "GOID"
) %>%
  dplyr::rename("GOALL" = GOID)

Annotationdbi_Annotated_ENTREZ_ENSEMBL_SYMBOL_GOALL <- Annotationdbi_Annotated_ENTREZ_ENSEMBL_SYMBOL %>%
  left_join(GOALL_ENTREZID,
    by = "ENTREZID",
    relationship = "many-to-many"
  ) %>%
  left_join(GO_Annotation,
    by = "GOALL"
  )


## Combine annotationdbi and biomart GO annotations
BiomaRt_Annotationdbi_GO <- Annotationdbi_Annotated_ENTREZ_ENSEMBL_SYMBOL_GOALL %>%
  dplyr::rename(
    "go_id" = GOALL,
    "name_1006" = TERM,
    "namespace_1003" = ONTOLOGYALL
  ) %>%
  dplyr::select(GENEID, go_id, name_1006, namespace_1003) %>%
  mutate(namespace_1003 = str_replace(namespace_1003, "BP", "biological_process")) %>%
  mutate(namespace_1003 = str_replace(namespace_1003, "MF", "molecular_function")) %>%
  mutate(namespace_1003 = str_replace(namespace_1003, "CC", "cellular_component")) %>%
  rbind(BiomaRt_ENSEMBL_Symbol_Entrez_GO_Annotation %>%
    dplyr::select(GENEID, go_id, name_1006, namespace_1003)) %>%
  dplyr::filter(!is.na(GENEID)) %>%
  dplyr::filter(!is.na(go_id)) %>%
  distinct() %>%
  full_join(genes_for_annotation,
    by = "GENEID"
  )

########################################## Annotate reactome gene sets with biomart
Reactome_gene <- getBM(
  mart = ensembl,
  attributes = c(
    "ensembl_gene_id",
    "reactome_gene"
  )
)

Reactome_gene_No_Empty <- Reactome_gene %>%
  dplyr::filter(reactome_gene != "") %>%
  distinct()

BiomaRt_Reactome_ENSEMBL_Symbol_Annotation <- GeneSymbol_biomaRt_Annotated_ENSEMBL_Symbol %>%
  full_join(Reactome_gene_No_Empty,
    by = "ensembl_gene_id",
    relationship = "many-to-many"
  ) %>%
  distinct()

Genes_Without_BiomaRt_Reactome_Annotation <- BiomaRt_Reactome_ENSEMBL_Symbol_Annotation[which(is.na(BiomaRt_Reactome_ENSEMBL_Symbol_Annotation$reactome_gene)), ]
# 37k genes lack GO-annotation in the ensembl database
sig_genes %>%
  dplyr::filter(GENEID %in% Genes_Without_BiomaRt_Reactome_Annotation$GENEID)
# significant genes lacking Reactome annotation

########################################## Annotate reactome with annotationdbi
keytypes(reactome.db)
Annotationdbi_Reactome_ENTREZ_ENSEMBL_SYMBOL <- as.data.frame(reactomeEXTID2PATHID) %>%
  dplyr::rename("ENTREZID" = gene_id) %>%
  dplyr::inner_join(Annotationdbi_ENTREZ_ENSEMBL_SYMBOL,
    by = "ENTREZID",
    relationship = "many-to-many"
  ) %>%
  dplyr::inner_join(as.data.frame(reactomePATHID2NAME),
    by = "DB_ID"
  )

BiomaRt_Reactome_Gene_Pairs <- BiomaRt_Reactome_ENSEMBL_Symbol_Annotation %>%
  dplyr::select(GENEID, reactome_gene) %>%
  drop_na() %>%
  distinct() %>%
  mutate(Reactome_Gene_Pair = paste(GENEID, reactome_gene, sep = "_"))

Annotationdbi_Reactome_Gene_Pairs <- Annotationdbi_Reactome_ENTREZ_ENSEMBL_SYMBOL %>%
  dplyr::select(ENSEMBL, DB_ID, path_name) %>%
  drop_na() %>%
  distinct() %>%
  mutate(Reactome_Gene_Pair = paste(ENSEMBL, DB_ID, sep = "_"))

Both_Reactome_Gene_Pairs <- BiomaRt_Reactome_Gene_Pairs %>%
  full_join(Annotationdbi_Reactome_Gene_Pairs,
    by = "Reactome_Gene_Pair"
  )

# Number of reactome mappings present in annotationdbi but not biomart
nrow(Both_Reactome_Gene_Pairs %>%
  dplyr::filter(is.na(GENEID)))

# Number of reactome mappings present in biomart but not annotationdbi
nrow(Both_Reactome_Gene_Pairs %>%
  dplyr::filter(is.na(ENSEMBL)))

# Reactome pathways in biomart but not annotation dbi
Both_Reactome_Gene_Pairs %>%
  dplyr::filter(!reactome_gene %in% Both_Reactome_Gene_Pairs$DB_ID) %>%
  dplyr::select(reactome_gene) %>%
  distinct()

# Reactome pathways in annotationdbi but not biomart
Both_Reactome_Gene_Pairs %>%
  dplyr::filter(!DB_ID %in% Both_Reactome_Gene_Pairs$reactome_gene) %>%
  dplyr::select(DB_ID) %>%
  distinct()

# Combine annotationdbi with biomart reactome annotation
BiomaRt_Annotationdbi_Reactome_ENTREZ_ENSEMBL_SYMBOL <- Annotationdbi_Reactome_ENTREZ_ENSEMBL_SYMBOL %>%
  dplyr::select(ENSEMBL, DB_ID) %>%
  dplyr::rename(
    "GENEID" = ENSEMBL,
    "reactome_gene" = DB_ID
  ) %>%
  rbind(BiomaRt_Reactome_ENSEMBL_Symbol_Annotation %>%
    dplyr::select(GENEID, reactome_gene)) %>%
  drop_na() %>%
  distinct() %>%
  full_join(
    Annotationdbi_Reactome_ENTREZ_ENSEMBL_SYMBOL %>%
      dplyr::rename("reactome_gene" = DB_ID) %>%
      dplyr::select(reactome_gene, path_name) %>%
      distinct(),
    by = "reactome_gene"
  ) %>%
  full_join(genes_for_annotation,
    by = "GENEID"
  ) %>%
  mutate(path_name = ifelse(is.na(path_name), reactome_gene, path_name))

########################################## KEGG annotation
## Get KEGG pathway ID to pathway name
KEGG_1 <- enframe(
  x = keggList(
    database = "pathway",
    organism = "mmu"
  ),
  name = "KEGG_ID",
  value = "KEGG_Name"
)
## Get KEGG pathway ID to ENTREZID
KEGG_2 <- enframe(
  x = keggLink(
    "mmu",
    "pathway"
  ),
  name = "KEGG_ID",
  value = "ENTREZID"
)
## Map KEGG pathway ID and name to ENTREZ id
KEGG_3 <- KEGG_2 %>%
  mutate(KEGG_ID = str_remove(
    KEGG_ID,
    "path:"
  )) %>%
  inner_join(KEGG_1,
    by = "KEGG_ID"
  ) %>%
  mutate(ENTREZID = str_remove(
    ENTREZID,
    "mmu:"
  ))

# Does our annotationdbi entrez annotation contain all entrez ids for the KEGG pathways?
nrow(KEGG_2) == nrow(KEGG_2 %>%
  mutate(ENTREZID = str_remove(
    ENTREZID,
    "mmu:"
  )) %>%
  dplyr::filter(ENTREZID %in% Annotationdbi_ENTREZ_ENSEMBL_SYMBOL$ENTREZID))

## Map KEGG to ENSEMBL and SYMBOL
KEGG_Annotated <- Annotationdbi_Annotated_ENTREZ_ENSEMBL_SYMBOL %>%
  full_join(KEGG_3,
    by = "ENTREZID",
    relationship = "many-to-many"
  )

########################################## Combine GO, Reactome and KEGG gene sets
annotationTable <- rbind(
  KEGG_Annotated %>%
    mutate(dbName = "KEGG") %>%
    dplyr::rename(c(
      "geneID" = GeneSymbol,
      "termID" = KEGG_ID,
      "termName" = KEGG_Name
    )) %>%
    dplyr::mutate(description = termName) %>%
    dplyr::select(geneID, termID, termName, description, dbName) %>%
    distinct() %>%
    dplyr::filter(!is.na(termID)) %>%
    dplyr::filter(!is.na(geneID)),
  BiomaRt_Annotationdbi_Reactome_ENTREZ_ENSEMBL_SYMBOL %>%
    mutate(dbName = "Reactome") %>%
    dplyr::rename(c(
      "geneID" = GeneSymbol,
      "termID" = reactome_gene,
      "termName" = path_name
    )) %>%
    dplyr::mutate(description = termName) %>%
    dplyr::select(geneID, termID, termName, description, dbName) %>%
    distinct() %>%
    dplyr::filter(!is.na(termID)) %>%
    dplyr::filter(!is.na(geneID)),
  BiomaRt_Annotationdbi_GO %>%
    dplyr::mutate(namespace_1003 = str_replace(namespace_1003,
                                               "^",
                                               "GO_")) %>% 
    dplyr::rename(c(
      "geneID" = GeneSymbol,
      "termID" = go_id,
      "termName" = name_1006,
      "dbName" = namespace_1003
    )) %>%
    dplyr::mutate(description = termName) %>%
    dplyr::select(geneID, termID, termName, description, dbName) %>%
    distinct() %>%
    dplyr::filter(!is.na(termID)) %>%
    dplyr::filter(!is.na(geneID))
) %>%
  dplyr::filter(dbName != "") %>%
  tidyr::unite("description", description:dbName, sep = " ", remove = FALSE)

# After doing some testing, it seems like including the description column like instructed in the package manual actually breaks the package
# Despite the script throwing warnings, it seems like description must NOT be included!
# This way, it uses the termname columns as description, despite the warnings of the column description not being present
annotationTable <- annotationTable %>% 
  dplyr::select(-description)

########################################## Prepare intermediate files, for example for GSEA with SetRank
## Export annotation dataframe, can help with reproducibility
write_tsv(
  x = annotationTable,
  file = "./R_intermediate_files/annotationTable.tsv"
)

## Create and export background reference gene set for SetRank
## The background will be all genes with at least 1 count in 1 sample
## When used by SetRank this should be a vector, but for exporting it's easier with 1-column table
referenceSet <- genes_for_annotation %>%
  dplyr::filter(Gene_Detected == TRUE) %>%
  dplyr::filter(GeneSymbol %in% annotationTable$geneID) %>%
  dplyr::select(GeneSymbol) %>%
  distinct() # %>%
# deframe()

write_csv(
  x = referenceSet,
  file = "./R_intermediate_files/referenceSet.tsv"
)

## Create and export gene identifiers for SetRank ranked by adjusted p-value
## When used by SetRank this should be a vector, but for exporting it's easier with 1-column table
geneIDs <- Combined_Results_DF %>%
  dplyr::filter(GeneSymbol %in% referenceSet$GeneSymbol) %>%
  dplyr::select(
    GeneSymbol,
    adj_pvalue
  ) %>%
  distinct() %>%
  arrange(adj_pvalue) %>%
  dplyr::select(GeneSymbol) # %>%
# deframe()

write_csv(
  x = geneIDs,
  file = "./R_intermediate_files/geneIDs.tsv"
)

## Export TPM genes with GO:0016757 glycosyltransferase activity (MF) annotation
# Long format
Glycosyltransferase_activity_genes_long <- BiomaRt_Annotationdbi_GO %>% 
  dplyr::filter(go_id == "GO:0016757") %>% 
  dplyr::filter(Gene_Detected == TRUE) %>% 
  dplyr::inner_join(TPM_df, by = "GeneSymbol") %>% 
  dplyr::filter(condition == "Non_Infected") %>% 
  dplyr::select(GeneSymbol, ensembldb_ID, TPM, User_ID, go_id) %>% 
  dplyr::distinct()

write_xlsx(x = Glycosyltransferase_activity_genes_long,
           path = "Glycosyltransferase_activity_genes_long.xlsx")

# Short format
Glycosyltransferase_activity_genes_wide <- BiomaRt_Annotationdbi_GO %>% 
  dplyr::filter(go_id == "GO:0016757") %>% 
  dplyr::filter(Gene_Detected == TRUE) %>% 
  dplyr::inner_join(TPM_df, by = "GeneSymbol") %>% 
  dplyr::filter(condition == "Non_Infected") %>% 
  dplyr::select(GeneSymbol, ensembldb_ID, TPM, User_ID, go_id) %>% 
  dplyr::distinct() %>% 
  dplyr::mutate(User_ID = str_c(User_ID,"_TPM")) %>% 
  pivot_wider(names_from = User_ID, values_from = TPM)

write_xlsx(x = Glycosyltransferase_activity_genes_wide,
           path = "Glycosyltransferase_activity_genes_wide.xlsx")

##################################################### Script finished
sessionInfo()

print("Script 2 finished")
